requirements.txt
comtypes==1.4.8
imbalanced_learn==0.12.3
matplotlib==3.9.2
mediapipe==0.10.18
numpy==2.1.3
pandas==2.2.3
Pillow==11.0.0
pyautogui==0.9.54
pycaw==20240210
roboflow==1.1.49
scikit_learn==1.5.2
seaborn==0.13.2
streamlit==1.37.1
tensorflow==2.18.0
tensorflow_intel==2.18.0
